Introduce new file structure for GPU deposition
This commit is meant to counter-act code cluttering by separating the different deposition routines into seperate files in a seperate folder. No changes to the functions themselves were made in this commit.
